Move content viewsets from urls.py into api.py
