+++ kolibri/content/api.py
@@ -1,0 +1,52 @@
+from django.db.models import Q
+from kolibri.content import models, serializers
+from rest_framework import filters, pagination, viewsets
+
+
+class ChannelMetadataCacheViewSet(viewsets.ModelViewSet):
+    serializer_class = serializers.ChannelMetadataCacheSerializer
+
+    def get_queryset(self):
+        return models.ChannelMetadataCache.objects.all()
+
+
+class ContentNodeFilter(filters.FilterSet):
+    search = filters.django_filters.MethodFilter(action='title_description_filter')
+
+    class Meta:
+        model = models.ContentNode
+        fields = ['parent', 'search', 'prerequisite_for', 'has_prerequisite', 'related']
+
+    def title_description_filter(self, queryset, value):
+        # only return the first 30 results to avoid major slow down
+        return queryset.filter(
+            Q(title__icontains=value) | Q(description__icontains=value)
+        )
+
+
+class OptionalPageNumberPagination(pagination.PageNumberPagination):
+    """
+    Pagination class that allows for page number-style pagination, when requested.
+    To activate, the `page_size` argument must be set. For example, to request the first 20 records:
+    `?page_size=20&page=1`
+    """
+    page_size = None
+    page_size_query_param = "page_size"
+
+
+class ContentNodeViewset(viewsets.ModelViewSet):
+    serializer_class = serializers.ContentNodeSerializer
+    filter_backends = (filters.DjangoFilterBackend,)
+    filter_class = ContentNodeFilter
+    pagination_class = OptionalPageNumberPagination
+
+    def get_queryset(self):
+        return models.ContentNode.objects.all()
+
+
+class FileViewset(viewsets.ModelViewSet):
+    serializer_class = serializers.FileSerializer
+    pagination_class = OptionalPageNumberPagination
+
+    def get_queryset(self):
+        return models.File.objects.all()
